Add files via upload
The ATM Simulation System is a console-based C program that mimics the basic operations of an Automated Teller Machine (ATM). It allows users to create accounts, log in using a PIN, check balance, deposit money, and withdraw money. The system stores user details (like name, PIN, and balance) in a text file, enabling data to persist even after the program is closed